SysMon: Work around m68k compiler issue with passing words to __sprintf()

git-svn-id: https://svn.aros.org/svn/aros/trunk@45236 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/system/SysMon/tasks.c
@@ -119,7 +119,7 @@ AROS_UFH3(VOID, TasksListDisplayFunction,
 
     if (obj)
     {
-        __sprintf(bufprio, "%d", obj->Priority);
+        __sprintf(bufprio, "%ld", (LONG)obj->Priority);
         strings[0] = obj->Name;
         strings[1] = bufprio;
         strings[2] = obj->Type == NT_TASK ? (STRPTR)_(MSG_TASK) : (STRPTR)_(MSG_PROCESS);
